fix: handle task callback correctly
The existing code would crash on calling the callback lambda for
a completed task because it takes `task` as parameter and our lambda
had no input. This change just passes discard that takes an element as
an argument. Discard is chosen over remove to ensure it won't raise
error (although it shouldn't happen).